Fix case issue with Navbar file
