+++ src/pages/Catalog/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import "../../app/globals.css";
 import { Navbar } from "@/components/Navbar";
-import { AppSidebar } from "@/components/sidebar"; // Import sidebar
+import { AppSidebar } from "@/components/sidebar"; 
 import Footer from "@/components/Footer";
 import ProductCard from "@/components/ProductCard";
 import { ProductPagination } from "@/components/pagination";
